A bunch of 64-bit compilation warning fixes. Implemented an array-based BVH tree in OcclusionManager.

+++ dependencies/Gaff/Gaff_Math.cpp
@@ -31,12 +31,12 @@ NS_GAFF
 	\param len The length of the buffer in bytes.
 	\return The calculated hash.
 */
-unsigned long long FNV1aHash64(const char* key, unsigned int len)
+unsigned long long FNV1aHash64(const char* key, size_t len)
 {
 	unsigned long long hash = 14695981039346656037ULL;
 
-	for (unsigned int i = 0; i < len; ++i) {
-		hash = (hash ^ (unsigned long long)key[i]) * 1099511628211ULL;
+	for (size_t i = 0; i < len; ++i) {
+		hash = (hash ^ static_cast<unsigned long long>(key[i])) * 1099511628211ULL;
 	}
 
 	return hash;
@@ -48,12 +48,12 @@ unsigned long long FNV1aHash64(const char* key, unsigned int len)
 	\param len The length of the buffer in bytes.
 	\return The calculated hash.
 */
-unsigned long long FNV1Hash64(const char* key, unsigned int len)
+unsigned long long FNV1Hash64(const char* key, size_t len)
 {
 	unsigned long long hash = 14695981039346656037ULL;
 
-	for (unsigned int i = 0; i < len; ++i) {
-		hash = (hash ^ 1099511628211ULL) * (unsigned long long)key[i];
+	for (size_t i = 0; i < len; ++i) {
+		hash = (hash ^ 1099511628211ULL) * static_cast<unsigned long long>(key[i]);
 	}
 
 	return hash;
@@ -65,12 +65,12 @@ unsigned long long FNV1Hash64(const char* key, unsigned int len)
 	\param len The length of the buffer in bytes.
 	\return The calculated hash.
 */
-unsigned int FNV1aHash32(const char* key, unsigned int len)
+unsigned int FNV1aHash32(const char* key, size_t len)
 {
 	unsigned int hash = 2166136261U;
 
-	for (unsigned int i = 0; i < len; ++i) {
-		hash = (hash ^ (unsigned int)key[i]) * 16777619U;
+	for (size_t i = 0; i < len; ++i) {
+		hash = (hash ^ static_cast<unsigned int>(key[i])) * 16777619U;
 	}
 
 	return hash;
@@ -82,12 +82,12 @@ unsigned int FNV1aHash32(const char* key, unsigned int len)
 	\param len The length of the buffer in bytes.
 	\return The calculated hash.
 */
-unsigned int FNV1Hash32(const char* key, unsigned int len)
+unsigned int FNV1Hash32(const char* key, size_t len)
 {
 	unsigned int hash = 2166136261U;
 
-	for (unsigned int i = 0; i < len; ++i) {
-		hash = (hash * 16777619U) ^ (unsigned int)key[i];
+	for (size_t i = 0; i < len; ++i) {
+		hash = (hash * 16777619U) ^ static_cast<unsigned int>(key[i]);
 	}
 
 	return hash;
